Middleware for data validation
