+++ controllers/product.controller.js
@@ -12,19 +12,6 @@ module.exports.create = (req, res) => {
 }
 
 module.exports.store = (req, res) => {
-	const errors = [];
-	if (!req.body.name) {
-		errors.push("Name is require");
-	} 
-	if (!req.body.price) {
-		errors.push("Price is require");
-	} 
-	if(!req.body.imgUrl) {
-		errors.push("image is require");
-	}
-	if (errors) {
-		res.render('admin/pages/add-product', { title: "Add product", errors: errors });
-	}
 
 	const name = req.body.name;
 	const price = req.body.price;
